fix(monitor): close metrics read/unmap race, fix v1 cache struct drift
collectPodAndContainerInfo read ContainerUsage.Info (backed by mmap'd
cache memory) without holding ContainerLister's mutex, while Update()
can concurrently munmap that same memory on its resync timer and
remove the entry - same use-after-unmap class as the loadCache fix
earlier in this branch, just on the read side. Wire the existing
Lock()/UnLock() (added earlier in this PR but never called anywhere)
around the read-and-scrape in collectPodAndContainerInfo.

Also fix v1.sharedRegionT's ownerPid field: it was typed uint32 while
libvgpu's C shared_region_t declares owner_pid as a size_t (8 bytes on
this cache format's 64-bit ABI). The two struct's total sizes matched
only because Go's own alignment padding ahead of the next uint64 field
happened to absorb the missing 4 bytes - every field after ownerPid
just happened to land on the right offset. Typing it uint64 makes the
layout match for the actual reason. Verified against libvgpu's real C
struct (multiprocess_memory_limit.h) with offsetof/sizeof on both
sides. Added TestSharedRegionTLayoutMatchesCABI, which checks every
field's offset (not just MinSize()'s total-size check) so a future
edit to either struct that breaks the layout fails loudly instead of
silently misreading vGPU accounting data.

+++ pkg/monitor/nvidia/v1/spec.go
@@ -59,10 +59,17 @@ type sharedRegionT struct {
 	majorVersion    int32
 	minorVersion    int32
 	smInitFlag      int32
-	ownerPid        uint32
-	sem             semT
-	num             uint64
-	uuids           [16]uuid
+	// ownerPid mirrors libvgpu's `_Atomic size_t owner_pid`, which is 8
+	// bytes on the 64-bit Linux ABI this cache format targets. It was
+	// previously typed uint32 (4 bytes); every field after it still lined
+	// up with the C layout only because Go's own 8-byte alignment padding
+	// ahead of `num` happened to absorb the missing 4 bytes. Typing it
+	// uint64 makes the two layouts match for the actual reason instead of
+	// by accident. See TestSharedRegionTLayoutMatchesCABI.
+	ownerPid uint64
+	sem      semT
+	num      uint64
+	uuids    [16]uuid
 
 	limit   [16]uint64
 	smLimit [16]uint64
